requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.1
python-dotenv>=1.0.0
strands-agents==1.28.0
strands-agents-tools==0.2.21
boto3==1.42.60
pydantic>=2.10.6
opensearch-py==3.1.0
requests-aws4auth==1.3.1
